move what belongs to GeneralReferralService to this class from generic OrchestrationService
