+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/service/impl/GeneralReferralServiceImplTest.java
@@ -12,6 +12,7 @@ import uk.gov.hmcts.reform.divorce.orchestration.domain.model.ccd.CcdCallbackRes
 import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.WorkflowException;
 import uk.gov.hmcts.reform.divorce.orchestration.service.CaseOrchestrationServiceException;
 import uk.gov.hmcts.reform.divorce.orchestration.workflows.generalreferral.GeneralConsiderationWorkflow;
+import uk.gov.hmcts.reform.divorce.orchestration.workflows.generalreferral.SetupGeneralReferralPaymentWorkflow;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -19,8 +20,14 @@ import java.util.Map;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.Assert.assertThrows;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_CASE_ID;
+import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_STATE;
+import static uk.gov.hmcts.reform.divorce.orchestration.controller.util.CallbackControllerTestUtils.assertCaseOrchestrationServiceExceptionIsSetProperly;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NO_VALUE;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.YES_VALUE;
 import static uk.gov.hmcts.reform.divorce.orchestration.testutil.GeneralReferralUtil.buildCallbackRequest;
@@ -31,6 +38,9 @@ public class GeneralReferralServiceImplTest {
 
     @Mock
     private GeneralConsiderationWorkflow generalConsiderationWorkflow;
+
+    @Mock
+    private SetupGeneralReferralPaymentWorkflow setupGeneralReferralPaymentWorkflow;
 
     @InjectMocks
     private GeneralReferralServiceImpl generalReferralService;
@@ -96,5 +106,34 @@ public class GeneralReferralServiceImplTest {
         when(generalConsiderationWorkflow.run(caseDetails)).thenThrow(WorkflowException.class);
 
         generalReferralService.generalConsideration(caseDetails);
+    }
+
+    @Test
+    public void givenCaseData_whenSetupConfirmServicePaymentWorkflow_thenReturnPayload() throws Exception {
+        CaseDetails caseDetails = CaseDetails.builder()
+            .caseData(new HashMap<>())
+            .caseId(TEST_CASE_ID)
+            .state(TEST_STATE)
+            .build();
+
+        when(setupGeneralReferralPaymentWorkflow.run(eq(caseDetails))).thenReturn(new HashMap<>());
+
+        generalReferralService.setupGeneralReferralPaymentEvent(caseDetails);
+
+        verify(setupGeneralReferralPaymentWorkflow).run(eq(caseDetails));
+    }
+
+    @Test
+    public void shouldThrowException_whenSetupGeneralReferralPaymentWorkflow_throwsWorkflowException() throws Exception {
+        when(setupGeneralReferralPaymentWorkflow.run(any())).thenThrow(WorkflowException.class);
+
+        CaseOrchestrationServiceException exception = assertThrows(
+            CaseOrchestrationServiceException.class,
+            () -> generalReferralService.setupGeneralReferralPaymentEvent(
+                CaseDetails.builder().caseId(TEST_CASE_ID).build()
+            )
+        );
+
+        assertCaseOrchestrationServiceExceptionIsSetProperly(exception);
     }
 }
